delete completed in all module and validation also completed

+++ app/Services/CategoryManagement.php
@@ -3,14 +3,18 @@
 namespace App\Services;
 
 use App\Repository\CategoryRepository ;
+use App\Repository\ProductRepository ;
 use Throwable;
 use Illuminate\Support\Facades\Log;
 
 class CategoryManagement
 {   
-    public function __construct(CategoryRepository $category_repo)
+    public $category_repo;
+    public $product_repo;
+    public function __construct(CategoryRepository $category_repo,ProductRepository $product_repo)
     {
         $this->category_repo = $category_repo;
+        $this->product_repo = $product_repo;
     }
 
     /**
@@ -30,8 +34,7 @@ class CategoryManagement
                 return ['result'=>'something went wrong'];
             }
         } catch (Throwable $e) {
-              Log::info($e->getMessage());
-            return view('error.error');
+            return  Log::error($e->getMessage());
         }
     }
 
@@ -52,8 +55,7 @@ class CategoryManagement
                return ['result'=>'something went wrong'];
            }
         } catch (Throwable $e) {
-              Log::info($e->getMessage());
-            return view('error.error');
+            return  Log::error($e->getMessage());
         }
     }
 
@@ -65,17 +67,23 @@ class CategoryManagement
     public function deleteRecord($id)
     {
         try {
-           $country=$this->category_repo->delete($id);
-           if($country)
-           {
-               return ['result'=>'record deleted'];
-           }
-           else{
-               return ['result'=>'something went wrong'];
-           }
+        $category_id=$this->product_repo->category_id($id);
+        if(!empty($category_id)){
+            return ['error'=>'id already assigned'];
+        }else{
+            $country=$this->category_repo->delete($id);
+            if($country)
+            {
+                return ['result'=>'record deleted'];
+            }else{
+                return ['result'=>'something went wrong'];
+            }
+        }
         } catch (Throwable $e) {
               Log::info($e->getMessage());
-            return view('error.error');
+            // return view('error.error');
+            return $e;
+        
         }
     }
